Setup improvement

+++ .devcontainer/setup.sh
@@ -1,8 +1,22 @@
 #!/bin/bash
 
-cp .env.example .env
+scriptPath=$(readlink -f "$0")
+scriptDir=$(dirname "$scriptPath")
+
+if ! [ -e "$scriptDir/../.env" ]; then
+  cp .env.example .env
+else
+  echo "File .env already exist"
+fi
+
 composer install
-php artisan key:generate
+
+if ! tr -d '\r' < .env | sed -n '/^APP_KEY=/s///p' | grep -q . ; then
+    php artisan key:generate
+else
+    echo "APP_KEY has a value assigned (php artisan key:generate ignored)"
+fi
+
 php artisan storage:link
 npm install
 npm run build
